Throw when saga mapping doesn't exist
