+++ knightbus/tests/KnightBus.Core.Tests.Unit/SagaHandlerTests.cs
@@ -1,4 +1,5 @@
 using System.Threading.Tasks;
+using FluentAssertions;
 using KnightBus.Core.Sagas;
 using KnightBus.Messages;
 using Moq;
@@ -21,7 +22,36 @@ namespace KnightBus.Core.Tests.Unit
             //act
             await handler.Initialize();
             //assert
-            store.Verify(x=> x.Create(id, It.IsAny<TestSagaData>()), Times.Once);
+            store.Verify(x => x.Create(id, It.IsAny<TestSagaData>()), Times.Once);
+        }
+
+        [Test]
+        public async Task Initialize_should_resume_saga_when_start_message_is_received()
+        {
+            //arrange
+            var id = "something-unique";
+            var store = new Mock<ISagaStore>();
+            var saga = new TestSaga();
+            var startMessage = new TestSagaMessage(id);
+            var handler = new SagaHandler<TestSagaData, TestSagaMessage>(store.Object, saga, startMessage);
+            saga.MessageMapper.MapMessage<TestSagaMessage>(x => x.MessageId);
+            //act
+            await handler.Initialize();
+            //assert
+            store.Verify(x => x.GetSaga<TestSagaData>(id), Times.Once);
+        }
+
+        [Test]
+        public void Initialize_should_throw_when_message_is_not_mapped()
+        {
+            //arrange
+            var id = "something-unique";
+            var store = new Mock<ISagaStore>();
+            var saga = new TestSaga();
+            var startMessage = new TestSagaMessage(id);
+            var handler = new SagaHandler<TestSagaData, TestSagaMessage>(store.Object, saga, startMessage);
+            //act & assert
+            handler.Awaiting(x => x.Initialize()).Should().Throw<SagaMessageMappingNotFoundException>();
         }
 
         internal class TestSagaData : ISagaData
@@ -41,7 +71,15 @@ namespace KnightBus.Core.Tests.Unit
             {
                 MessageId = id;
             }
-            public string MessageId { get;} 
+            public string MessageId { get; }
+        }
+        internal class TestSagaMessage : IMessage
+        {
+            public TestSagaMessage(string id)
+            {
+                MessageId = id;
+            }
+            public string MessageId { get; }
         }
     }
 }
